Add Bullet class and shoot fuction
MM now is able to shoot bullets. When the user press the shoot key, the
createBullet() method is being called from the BulletListener interface.
The Level implements the class and sets the listener on MegaMan. Every
bullet which left the screen are being set to inactive. The Level
removes the inactive bullets which are being stored in an array.