Add mobile controls
Add MobileControls class and MobileControlListener interface. It's int
eh fist phase, with the left and the right button, represented by
rectangles. The tricky part was to figure out how to create the camera
and viewport same size as the WORLD_SIZE. The input polling was also
tricky, and the communication between